requirements.txt
Pillow==9.4.0  
paddlepaddle  
paddleocr>=2.8.1 
colorama
fastapi
uvicorn
requests
schedule
